token http only cookie confirmation and readme.md file

+++ routers/tools.ts
@@ -2,6 +2,7 @@ import { Router } from 'express';
 import { ToolsRecord } from '../database/tool.record';
 import { HistoryRecord } from '../database/history.record';
 import { handleError } from '../utils/error';
+import { checkToken } from '../utils/loginConfirm';
 
 const toolRouter = Router();
 
@@ -9,8 +10,15 @@ toolRouter
 
     .get('/', async (req, res, next) => {
         try {
-            const data = await ToolsRecord.getAll();
-            res.json(data);
+            const token = req.cookies.token;
+            const confirmation = checkToken(token);
+            if(confirmation) {
+                const data = await ToolsRecord.getAll();
+                res.json(data);
+            } else {
+                res.json({error: 'Brak potwierdzenia zalogowania: Brak dostępu'})
+            }
+            
         } catch (err) {
             handleError(err, req, res, next);
         }
@@ -19,8 +27,15 @@ toolRouter
     .get('/:id', async (req, res, next) => {
         const { id } = req.params;
         try {
-            const data = await ToolsRecord.getOne(id);
-            res.json(data);
+            const token = req.cookies.token;
+            const confirmation = checkToken(token);
+            if(confirmation) {
+                const data = await ToolsRecord.getOne(id);
+                res.json(data);
+            } else {
+                res.json({error: 'Brak potwierdzenia zalogowania: Brak dostępu'})
+            }
+            
         } catch (err) {
             handleError(err, req, res, next);
         }
@@ -29,11 +44,18 @@ toolRouter
     .post('/', async (req, res, next) => {
         const { sign, name, status, place, type, subtype, brand, serial } = req.body;
         try {
-            const tool = new ToolsRecord({ sign, name, status, place, type, subtype, brand, serial });
-            const id = await tool.add();
-            const history = new HistoryRecord({ id, name });
-            await history.add();
-            res.json({ id });
+            const token = req.cookies.token;
+            const confirmation = checkToken(token);
+            if(confirmation) {
+                const tool = new ToolsRecord({ sign, name, status, place, type, subtype, brand, serial });
+                const id = await tool.add();
+                const history = new HistoryRecord({ id, name });
+                await history.add();
+                res.json({ id });
+            } else {
+                res.json({error: 'Brak potwierdzenia zalogowania: Brak dostępu'})
+            }
+            
         } catch (err) {
             handleError(err, req, res, next);
         }
@@ -42,9 +64,16 @@ toolRouter
     .delete('/', async (req, res, next) => {
         const { id } = req.body;
         try {
-            const tool = await ToolsRecord.getOne(id);
-            const idn = await tool.delete();
-            res.json({ id: idn });
+            const token = req.cookies.token;
+            const confirmation = checkToken(token);
+            if(confirmation) {
+                const tool = await ToolsRecord.getOne(id);
+                const idn = await tool.delete();
+                res.json({ id: idn });
+            } else {
+                res.json({error: 'Brak potwierdzenia zalogowania: Brak dostępu'})
+            }
+            
         } catch (err) {
             handleError(err, req, res, next);
         }
@@ -53,16 +82,23 @@ toolRouter
     .patch('/', async (req, res, next) => {
         const { id, name, status, place } = req.body;
         try {
-            const actualizeRecord = await ToolsRecord.getOne(id);
-            const oldName = actualizeRecord.showStatuses.name;
-            const uuid = await actualizeRecord.actualize({ name, status, place });
-            if (oldName !== name) {
-                const newWorker = new HistoryRecord({ id, name });
-                await newWorker.add();
-                const oldWorker = await HistoryRecord.getOne(uuid.uuid);
-                await oldWorker.addEnd();
-            }
+            const token = req.cookies.token;
+            const confirmation = checkToken(token);
+            if(confirmation) {
+                const actualizeRecord = await ToolsRecord.getOne(id);
+                const oldName = actualizeRecord.showStatuses.name;
+                const uuid = await actualizeRecord.actualize({ name, status, place });
+                if (oldName !== name) {
+                    const newWorker = new HistoryRecord({ id, name });
+                    await newWorker.add();
+                    const oldWorker = await HistoryRecord.getOne(uuid.uuid);
+                    await oldWorker.addEnd();
+                }
             res.json({ id })
+            } else {
+                res.json({error: 'Brak potwierdzenia zalogowania: Brak dostępu'})
+            }
+            
         } catch (err) {
             handleError(err, req, res, next);
         }
